Update: add comments and clear to pass the norm
